Move functional test cases to their own module
In preparation of clean-up of the tests.
